LayoutTests/fast/events/spatial-navigation/snav-div-scrollable-but-without-focusable-content.html

<html>
  <!--
    This test ensures the correctness of a basic aspect of spatial navigation traversal
    routines: scrollable block html elements without any focusable content should be ignored.

    * Pre-conditions:
    1) DRT support for spatial navigation enable/disable.

    * Navigation steps:
    1) Loads this page, focus goes to "start" automatically.
    2) Focus moves from "start" to "end" bypassing the two div's
       in between them, once both have no focusable elements inside.
  -->
  <head>
    <style type="text/css">
      div.scroll {
        height: 200px;
        width: 300px;
        overflow: auto;
        border: 1px solid #666;
        background-color: #ccc;
        padding: 8px;
      }
    </style>

    <script src="../../js/resources/js-test-pre.js"></script>
    <script src="resources/spatial-navigation-utils.js"></script>
    <script type="application/javascript">

    var resultMap = [
      ["Down", "end"],
      ["Up"  , "start"],
      ["DONE", "DONE"]
    ];

    if (window.layoutTestController) {
      layoutTestController.dumpAsText();
      layoutTestController.setSpatialNavigationEnabled(true);
      layoutTestController.overridePreference("WebKitTabToLinksPreferenceKey", 1);
      layoutTestController.waitUntilDone();
    }

    function runTest()
    {
      // starting the test itself: get to a known place.
      document.getElementById("start").focus();

      initTest(resultMap, testCompleted);
    }

    function testCompleted()
    {
      if (window.layoutTestController)
        layoutTestController.notifyDone();
    }

    window.onload = runTest;
    </script>
    <script src="../js/resources/js-test-post.js"></script>
  </head>
  <body id="some-content" xmlns="http://www.w3.org/1999/xhtml">
    <div><a id="start" href="a">a</a></div>
    <div class="scroll">
      <p>This is a scrollable Div created with the CSS property overflow.</p>
      <p>It has no keyboard focusable elements ....</p>
      <br><br><br><br><br><br>
      <br><br><br><br><br><br>
      <p>... and scrollbars!</p>
    </div>

    <div class="scroll">
      <p>This is another scrollable div created with the CSS property overflow.</p>
      <p>It has also no keyboard focusable elements ....</p>
      <br><br><br><br><br><br>
      <p>... as well as scrollbars!</p>
    </div>
    <div><a id="end" href="a">e</a></div>
    <div id="console"></div>
  </body>
</html>
