LayoutTests/fast/events/spatial-navigation/snav-div-scrollable-but-without-focusable-content-expected.txt

a
This is a scrollable Div created with the CSS property overflow.

It has no keyboard focusable elements ....













... and scrollbars!

This is another scrollable div created with the CSS property overflow.

It has also no keyboard focusable elements ....







... as well as scrollbars!

e
PASS document.activeElement.getAttribute("id") is "end"
PASS document.activeElement.getAttribute("id") is "start"
